memory leak and a couple of warnings
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/SkinnedMesh@774 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CB3DMeshLoader.cpp
@@ -719,19 +719,19 @@ bool CB3DMeshLoader::readChunkKEYS(CSkinnedMesh::SJoint *InJoint)
 		if (flags & 1)
 		{
 			CSkinnedMesh::SPositionKey *Key=AnimatedMesh->createPositionKey(InJoint);
-			Key->frame = frame;
+			Key->frame = (f32)frame;
 			Key->position = position;
 		}
 		if (flags & 2)
 		{
 			CSkinnedMesh::SScaleKey *Key=AnimatedMesh->createScaleKey(InJoint);
-			Key->frame = frame;
+			Key->frame = (f32)frame;
 			Key->scale=scale;
 		}
 		if (flags & 4)
 		{
 			CSkinnedMesh::SRotationKey *Key=AnimatedMesh->createRotationKey(InJoint);
-			Key->frame = frame;
+			Key->frame = (f32)frame;
 			Key->rotation = rotation;
 		}
 	}
